[libc-trusty] Fixup for 64-bit apps
Bug: 120497636

+++ lib/libc-trusty/malloc.c
@@ -61,7 +61,7 @@ static void* sbrk(ptrdiff_t increment) {
     start = (char*)ROUNDUP((long)__libc_brk, SBRK_ALIGN);
     end = start + ROUNDUP((long)increment, SBRK_ALIGN);
 
-    new_brk = (char*)brk((uint32_t)end);
+    new_brk = (char*)brk((uint32_t)(uintptr_t)end);
     if (new_brk < end)
         return (void*)-1;
